[CI] Remove hardcoded `opensearch.version` from CI "Build and Tun Tests" job
OpenSearch version is typically specified in `build.gradle` file (or in `gradle.properties` file which is currently not present in this template). However, as of writing the CI job for this template silently overrides this value to whatever is hardcoded in `CI.yml` file. This leads to two issues:

- when the template code is upgraded to a new version of OpenSearch the version needs to be changed at least in two places (which is prone to errors)
- if the plugin author is using GHA then most likely he/she is not aware of the fact that the CI job is overriding the version (a missing docs is to blame?) and they can face unexpected failures of the job

If the `opensearch.version` needs to be overridden for the CI job I suggest documenting it properly and giving plugin authors instruction on **how** and **when** to modify the version in CI.yml file.

Closes #21

Signed-off-by: Lukáš Vlček <lukas.vlcek@aiven.io>