Clarify plugin repo naming conventions
OpenSearch plugins can "live" either inside the OpenSearch organization or outside it. This is the difference between official plugins and 3rd party plugins that are maintained in independent repositories.

Currently, these two types of plugins are not recommended to share exactly the same repo naming conventions (this might change but not in the near-term future). If needed, the 3rd party plugin authors should also follow up with several extra checks in making sure the formal processes are correctly aligned with the community needs (such as Code of Conduct, Security Policy, Trademarks and Attributions).

Closes #23

Signed-off-by: Lukáš Vlček <lukas.vlcek@aiven.io>