enable dependabot for java + GHA
this will enable dependabot updates, as it is also done for other
OpenSearch projects.
while this project currently doesn't have any java dependencies it still
makes sense to enable it here since this is a template repository and
projects created based on it will usually have such dependencies in the
future.

Signed-off-by: Ralph Ursprung <Ralph.Ursprung@avaloq.com>