CI: run gradle commands in separate steps
right now `./gradlew build` could fail but the script would still go on
(because neither `set -e` is set nor `&&` is used to run both commands
together).

running it in two steps gives more visibilty to which step is failing.
note that `publishPluginZipPublicationToZipStagingRepository` is tested
explicitly to ensure that this step works as it has previously failed.

fixes #60

Signed-off-by: Ralph Ursprung <Ralph.Ursprung@avaloq.com>