remove manual DCO check (done by org-level check)
the DCO check is activated directly on GitHub in the org-level settings
for `opensearch-project` (see the [GitHub documentation]). there's no
need for the duplicate check implemented in the workflow anymore.

other projects which are created based on the template which want a DCO
check can also just enable this in their settings.

fixes #61

[GitHub documentation]: https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/managing-repository-settings/managing-the-commit-signoff-policy-for-your-repository

Signed-off-by: Ralph Ursprung <Ralph.Ursprung@avaloq.com>